altera position, add keyboard, altera mapa e Makefile
Position deve trabalhar com floats e nao ints; devemos ter metodos de update para cada modelo; nao faz sentido ter a Nave na queue de cenario ja que esta nunca sairia da mesma; precisa-se iniciar a lib de keyboard

+++ src/lib/position.c
@@ -1,6 +1,6 @@
 #include "position.h"
 
-Position POS_create(int x, int y, int z)
+Position POS_create(float x, float y, float z)
 {
 	Position *p = malloc(sizeof *p);
